./video_editor.py

+++ VideoEditor/video_editor_gui.py
@@ -1,56 +1,54 @@
+#!/usr/bin/python3
+
 import os
 import tkinter as tk
 from tkinter import filedialog, messagebox, ttk
-import subprocess
+import sys
+import time
+import video_editor
 
 curr_directory = os.getcwd()
 process_output_success = "The edited video can be found at {}/output_video.mp4".format(curr_directory)
 process_output_failure = "There was an error. Contact Brad"
-TIME_LABEL_TEXT = "Time (duration of silence in between non-silence to be cutout from video (seconds) [default: 2]):"
+TIME_LABEL_TEXT = "Time (duration of silence in between non-silence to be cut out from video (seconds) [default: 2]):"
 THRESHOLD_LABEL_TEXT = "Threshold (manually adjust silence threshold, which determines minimum sound level to be considered silence [default: 1500]):"
-OUTPUT_FILE_LOCATION_TEXT = "File location, use 'Browse' to naviagate through your file navigation system (location of video file [default: None])"
-
-
-def run_script():
-    time = time_entry.get()
-    threshold = threshold_entry.get()
-    file_path = file_entry.get()
-
-    if not os.path.exists(file_path):
-        messagebox.showerror("Error", "File does not exist!")
-        return
-
-    command = ['python', '{}/video_editor.py'.format(curr_directory), time, threshold, file_path]
-
-    # Run the script using subprocess.Popen
-    global process
-    process = subprocess.Popen(command)
-
-    update_progress()
-
-def update_progress():
-    return_code = process.poll()
-    if return_code is None:
-        progress["value"] += 1
-        if progress["value"] >= progress["maximum"]:
-            progress["value"] = 0
-        progress.after(100, update_progress)
-    else:
-        progress["value"] = progress["maximum"]
-        if return_code == 0:
-            messagebox.showinfo("Finished Editing!", process_output_success)
-        else:
-            messagebox.showinfo("Error", process_output_failure)
+OUTPUT_FILE_LOCATION_TEXT = "File location, use 'Browse' to navigate through your file navigation system (location of video file [default: None])"
 
 def browse_file():
-    file_path = filedialog.askopenfilename(initialdir="/", title="Select File")
-    file_entry.delete(0, tk.END)
-    file_entry.insert(tk.END, file_path)
+    filepath = filedialog.askopenfilename(initialdir=curr_directory, title="Select Video File",
+                                          filetypes=(("MP4 files", "*.mp4"), ("all files", "*.*")))
+    if filepath:
+        file_entry.delete(0, tk.END)
+        file_entry.insert(tk.END, filepath)
 
-def start_script():
-    start_button["state"] = tk.DISABLED
-    run_script()
-    start_button["state"] = tk.NORMAL
+
+def update_progress(progress):
+    progress_bar["value"] = progress
+    window.update_idletasks()
+
+
+def run_video_editor():
+    filepath = file_entry.get()
+    if not filepath:
+        messagebox.showwarning("Error", "Please select a video file.")
+        return
+
+    time_value = time_entry.get()
+    if not time_value:
+        time_value = 2
+
+    threshold_value = threshold_entry.get()
+    if not threshold_value:
+        threshold_value = 1500
+
+    progress_bar["value"] = 0
+    progress_bar["maximum"] = 100
+
+    # Call the video_editor.main function with the progress callback
+    video_editor.main(int(time_value), int(threshold_value), filepath, progress_callback=update_progress)
+
+    messagebox.showinfo("Video Editor", "Video editing complete.")
+
 
 def check_label_content(*args):
     if file_entry.get():
@@ -58,7 +56,7 @@ def check_label_content(*args):
     else:
         start_button.config(state=tk.DISABLED)
 
-# Create the main window
+
 window = tk.Tk()
 window.title("Brad's Video Editor")
 window.configure(bg="white")
@@ -91,10 +89,10 @@ time_entry.pack(pady=(5, 0))
 threshold_label = ttk.Label(window, text=THRESHOLD_LABEL_TEXT)
 threshold_label.pack(pady=(30, 0))
 threshold_entry = ttk.Entry(window, width=40)
-threshold_entry.pack(pady=(5,0))
+threshold_entry.pack(pady=(5, 0))
 
 browse_file_label = ttk.Label(window, text=OUTPUT_FILE_LOCATION_TEXT)
-browse_file_label.pack(pady=(30,0))
+browse_file_label.pack(pady=(30, 0))
 
 label_var = tk.StringVar()
 label_var.trace("w", check_label_content)
@@ -108,11 +106,11 @@ browse_button.pack(pady=(5, 0))
 
 progress_label = ttk.Label(window, text="Progress Bar")
 progress_label.pack(pady=(50, 0))
-progress = ttk.Progressbar(window, length=300, mode="determinate")
-progress.pack(pady=10)
+progress_bar = ttk.Progressbar(window, length=300, mode="determinate")
+progress_bar.pack(pady=10)
 
 # Create a button to run the script
-start_button = ttk.Button(window, text="Start", command=start_script, state=tk.DISABLED)
+start_button = ttk.Button(window, text="Start", command=run_video_editor, state=tk.DISABLED)
 start_button.pack(pady=30)
 
 # Start the GUI event loop
